refactor(tests): extract FakeUserRepository to sibling module
