+++ tests/unit/api/fakes.py
@@ -4,9 +4,7 @@ import asyncio
 from datetime import datetime
 from typing import Any
 
-from synthorg.api.auth.models import ApiKey, OrgRole, User
-from synthorg.api.auth.system_user import is_system_user
-from synthorg.api.guards import HumanRole
+from synthorg.api.auth.models import ApiKey
 from synthorg.budget.cost_record import CostRecord
 from synthorg.communication.channel import Channel
 from synthorg.communication.message import Message
@@ -29,17 +27,7 @@ from synthorg.hr.performance.models import (
     CollaborationMetricRecord,
     TaskMetricRecord,
 )
-from synthorg.persistence.constraint_tokens import (
-    IDX_SINGLE_CEO,
-    LAST_CEO_TRIGGER,
-    LAST_OWNER_TRIGGER,
-    USERS_USERNAME_UNIQUE,
-)
-from synthorg.persistence.errors import (
-    ConstraintViolationError,
-    DuplicateRecordError,
-    QueryError,
-)
+from synthorg.persistence.errors import DuplicateRecordError, QueryError
 from synthorg.persistence.preset_repository import PresetListRow, PresetRow
 from synthorg.security.models import AuditEntry, AuditVerdictStr
 from synthorg.security.timeout.parked_context import ParkedContext
@@ -300,124 +288,6 @@ class FakeAuditRepository:
 from tests.unit.api.fakes_decisions import (  # noqa: E402
     FakeDecisionRepository as FakeDecisionRepository,  # noqa: PLC0414
 )
-
-
-class FakeUserRepository:
-    """In-memory user repository for tests.
-
-    Enforces the same constraints as the real DB:
-    - Unique username
-    - At most one CEO (unique partial index)
-    - At least one CEO (trigger on role change)
-    - At least one owner (trigger on org_roles change)
-    """
-
-    def __init__(self) -> None:
-        self._users: dict[str, User] = {}
-
-    async def save(self, user: User) -> None:
-        existing = self._users.get(user.id)
-        # Username uniqueness
-        for u in self._users.values():
-            if u.username == user.username and u.id != user.id:
-                msg = "UNIQUE constraint failed: users.username"
-                raise ConstraintViolationError(msg, constraint=USERS_USERNAME_UNIQUE)
-        # CEO uniqueness (partial unique index on role='ceo')
-        if user.role == HumanRole.CEO:
-            for u in self._users.values():
-                if u.role == HumanRole.CEO and u.id != user.id:
-                    msg = "UNIQUE constraint failed: idx_single_ceo"
-                    raise ConstraintViolationError(
-                        msg,
-                        constraint=IDX_SINGLE_CEO,
-                    )
-        # Last-CEO trigger: prevent demoting the only CEO
-        if (
-            existing is not None
-            and existing.role == HumanRole.CEO
-            and user.role != HumanRole.CEO
-        ):
-            other_ceos = sum(
-                1
-                for u in self._users.values()
-                if u.role == HumanRole.CEO and u.id != user.id
-            )
-            if other_ceos == 0:
-                msg = "Cannot remove the last CEO"
-                raise ConstraintViolationError(
-                    msg,
-                    constraint=LAST_CEO_TRIGGER,
-                )
-        # Last-owner trigger: prevent removing the last owner
-        if (
-            existing is not None
-            and OrgRole.OWNER in existing.org_roles
-            and OrgRole.OWNER not in user.org_roles
-        ):
-            other_owners = sum(
-                1
-                for u in self._users.values()
-                if u.id != user.id and OrgRole.OWNER in u.org_roles
-            )
-            if other_owners == 0:
-                msg = "Cannot remove the last owner"
-                raise ConstraintViolationError(
-                    msg,
-                    constraint=LAST_OWNER_TRIGGER,
-                )
-        self._users[user.id] = user
-
-    async def get(self, user_id: str) -> User | None:
-        return self._users.get(user_id)
-
-    async def get_by_username(self, username: str) -> User | None:
-        for user in self._users.values():
-            if user.username == username:
-                return user
-        return None
-
-    async def list_users(self) -> tuple[User, ...]:
-        return tuple(u for u in self._users.values() if u.role != HumanRole.SYSTEM)
-
-    async def count(self) -> int:
-        return sum(1 for u in self._users.values() if u.role != HumanRole.SYSTEM)
-
-    async def count_by_role(self, role: HumanRole) -> int:
-        return sum(1 for u in self._users.values() if u.role == role)
-
-    async def delete(self, user_id: str) -> bool:
-        if is_system_user(user_id):
-            msg = "System user cannot be deleted"
-            raise QueryError(msg)
-        user = self._users.get(user_id)
-        if user is None:
-            return False
-        if user.role == HumanRole.CEO:
-            other_ceos = sum(
-                1
-                for u in self._users.values()
-                if u.role == HumanRole.CEO and u.id != user_id
-            )
-            if other_ceos == 0:
-                msg = "Cannot remove the last CEO"
-                raise ConstraintViolationError(
-                    msg,
-                    constraint=LAST_CEO_TRIGGER,
-                )
-        if OrgRole.OWNER in user.org_roles:
-            other_owners = sum(
-                1
-                for u in self._users.values()
-                if u.id != user_id and OrgRole.OWNER in u.org_roles
-            )
-            if other_owners == 0:
-                msg = "Cannot remove the last owner"
-                raise ConstraintViolationError(
-                    msg,
-                    constraint=LAST_OWNER_TRIGGER,
-                )
-        del self._users[user_id]
-        return True
 
 
 class FakeApiKeyRepository:
@@ -711,6 +581,32 @@ class FakeSettingsRepository:
             **self._store,
             (namespace, key): (value, updated_at),
         }
+        return True
+
+    async def set_many(
+        self,
+        items: Any,
+        *,
+        expected_updated_at_map: Any = None,
+    ) -> bool:
+        if not items:
+            return True
+        cas_map = expected_updated_at_map or {}
+        snapshot = dict(self._store)
+        draft = dict(self._store)
+        for namespace, key, value, updated_at in items:
+            expected = cas_map.get((namespace, key))
+            if expected is not None:
+                current = draft.get((namespace, key))
+                if current is None:
+                    if expected != "":
+                        self._store = snapshot
+                        return False
+                elif current[1] != expected:
+                    self._store = snapshot
+                    return False
+            draft[(namespace, key)] = (value, updated_at)
+        self._store = draft
         return True
 
     async def delete(self, namespace: str, key: str) -> bool:
